# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vigilancia - Porterias APP API.",
    "version": "1.0.1",
    "description": "Documentación de Vigilancias - Porterias API"
  },
  "paths": {
...
    "/api/v1/auth/verifyUser": {
      "post": {
        "summary": "Verifica un usuario por cuenta",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyUserByAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario verificado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyUserByAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyUserByAccountResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/authToken": {
      "get": {
        "summary": "Verifica un token de autenticación",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token verificado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyAuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token expirado o falta el encabezado de autorización",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "VerifyUserByAccountRequest": {
        "type": "object",
        "required": [
          "phone",
          "account_id"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "description": "Número de teléfono del usuario"
          },
          "account_id": {
            "type": "integer",
            "description": "ID de la cuenta asociada al usuario",
            "format": "int64"
          }
        }
      },
      "VerifyUserByAccountResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Código de estado de la respuesta"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la respuesta"
          },
          "token": {
            "type": "string",
            "nullable": true,
            "description": "Token de autenticación (puede ser nulo)"
          }
        },
        "required": [
          "status",
          "message",
          "token"
        ]
      },
      "VerifyAuthTokenResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "Código de estado de la respuesta"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la respuesta"
          },
          "data": {
            "$ref": "#/components/schemas/UserAttributes",
            "nullable": true,
            "description": "Atributos del usuario (puede ser nulo)"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
      "UserAttributes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "description": "Atributos del usuario"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
}